msm: kgsl: Delay 30ms to make sure GPU is ready during resume time
During resume, the kgsl driver preforms a soft reset of the GPU and it
needs to wait certain amount of time for GPU to be ready.
On most chips 10ms is sufficient. But on 8660 with newer RPM images,
the GPU often hangs with the 10ms delay, but testing has shown that
30ms does not hang.

CRs-Fixed: 286145
Signed-off-by: Wei Zou <wzou@codeaurora.org>
(cherry picked from commit d1f7f808d72a5049a8b91c30282a9050e467c74d)

Change-Id: I8c1b97587e35eca36341885c88bdb44c52f33701
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>